refactor(v2): extract all v2 mappers into hooks/v2/mappers/
