+++ src/hooks/v2/useVotesV2.ts
@@ -1,26 +1,12 @@
 import { useQuery } from '@tanstack/react-query';
 import { apolloClientQuery } from '../../lib/apolloClient';
 import { Vote } from '../../graphql/subgraph';
+import { ClassicVoteV2Profile, mapClassicVoteV2ToVote } from './mappers/vote';
 
 interface Props {
   chainId: string;
   jurorID?: string;
   enabled?: boolean;
-}
-
-interface ClassicVoteV2 {
-  id: string;
-  choice: string;
-  voted: boolean;
-  coreDispute: {
-    id: string;
-    period: string;
-    currentRuling: string;
-    templateId: string | null;
-    lastPeriodChange: string;
-    court: { id: string };
-  };
-  localRound: { id: string };
 }
 
 const CLASSIC_VOTES_QUERY = `
@@ -48,30 +34,6 @@ const CLASSIC_VOTES_QUERY = `
   }
 `;
 
-function mapClassicVoteV2ToVote(v: ClassicVoteV2): Vote {
-  // localRound.id format: "{kitId}-{disputeId}-{roundIndex}"
-  const roundIndex = v.localRound.id.split('-').at(-1) ?? '0';
-  const roundId = `${v.coreDispute.id}-${roundIndex}`;
-
-  return {
-    id: v.id,
-    voted: v.voted,
-    choice: v.choice,
-    round: { id: roundId },
-    address: { id: '' },
-    dispute: {
-      id: v.coreDispute.id,
-      currentRulling: v.coreDispute.currentRuling,
-      subcourtID: { id: v.coreDispute.court.id },
-      period: v.coreDispute.period,
-      arbitrable: { id: '' },
-      templateId: v.coreDispute.templateId,
-    },
-    // v2 has no exact vote timestamp — use lastPeriodChange as approximation
-    timestamp: v.coreDispute.lastPeriodChange,
-  };
-}
-
 export const useVotesV2 = ({ chainId, jurorID, enabled = true }: Props) => {
   return useQuery<Vote[], Error>({
     queryKey: ['useVotesV2', chainId, jurorID],
@@ -82,11 +44,15 @@ export const useVotesV2 = ({ chainId, jurorID, enabled = true }: Props) => {
 
       // Paginate through all votes
       while (true) {
-        const response = await apolloClientQuery<{ classicVotes: ClassicVoteV2[] }>(chainId, CLASSIC_VOTES_QUERY, {
-          jurorId: jurorID!.toLowerCase(),
-          first: 1000,
-          skip: skip,
-        });
+        const response = await apolloClientQuery<{ classicVotes: ClassicVoteV2Profile[] }>(
+          chainId,
+          CLASSIC_VOTES_QUERY,
+          {
+            jurorId: jurorID!.toLowerCase(),
+            first: 1000,
+            skip: skip,
+          },
+        );
 
         if (!response || !response.data) throw new Error('No response from TheGraph');
